requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-redis==5.4.0
celery==5.3.6
eventlet==0.35.1
qiniu==7.13.0